Rename to profile_grids and update README
